requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Google Drive and Sheets integration
pydrive2>=1.15.0
pydrive>=1.3.1
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
gspread>=5.0.0
oauth2client>=4.1.3

# Document processing
python-docx>=0.8.11
PyMuPDF>=1.23.0
docx2txt>=0.8
pypdf>=3.0.0
pdfminer.six>=20220524

# AI/ML and API clients
openai>=1.0.0
requests>=2.28.0

# Utilities and progress tracking
tqdm>=4.64.0
python-dotenv>=0.19.0
tenacity>=8.0.0
rapidfuzz>=3.6.1

# Excel and spreadsheet processing
openpyxl>=3.0.0

# Additional utilities
pathlib2>=2.3.0
typing-extensions>=4.0.0

# Vercel specific
vercel>=0.0.0